Fixed jumping bug in auto-size textareas.
How you ask? With a wrapper div of course. How else?
Now the max height is 95vh (95% of viewport height)
When a textarea is resized, the container div is set to
the current height, to avoid the content flow jitter.
Once the height is set, the container is restored to auto-height.